+++ concordium-sdk/src/main/java/com/concordium/sdk/AccountRequest.java
@@ -2,6 +2,7 @@ package com.concordium.sdk;
 
 import com.concordium.sdk.responses.AccountIndex;
 import com.concordium.sdk.transactions.AccountAddress;
+import com.concordium.sdk.transactions.CredentialRegistrationId;
 import com.concordium.sdk.transactions.Hash;
 import com.google.protobuf.ByteString;
 
@@ -16,11 +17,13 @@ public final class AccountRequest {
     private final Type type;
     private final AccountAddress address;
     private final AccountIndex index;
+    private final CredentialRegistrationId registrationId;
 
-    private AccountRequest(Type type, AccountAddress address, AccountIndex index) {
+    private AccountRequest(Type type, AccountAddress address, AccountIndex index, CredentialRegistrationId credentialRegistrationId) {
         this.type = type;
         this.address = address;
         this.index = index;
+        this.registrationId = credentialRegistrationId;
     }
 
     ByteString getByteString() {
@@ -29,6 +32,8 @@ public final class AccountRequest {
                 return ByteString.copyFrom(address.getEncodedBytes());
             case INDEX:
                 return ByteString.copyFrom(Long.toString(index.getIndex().getValue()).getBytes(StandardCharsets.UTF_8));
+            case CREDENTIAL_REGISTRATION_ID:
+                return ByteString.copyFromUtf8(registrationId.getEncoded());
             default:
                 throw new IllegalStateException("Invalid AccountRequest Type " + type);
         }
@@ -41,7 +46,7 @@ public final class AccountRequest {
      * @return the AccountRequest
      */
     public static AccountRequest from(AccountAddress address) {
-        return new AccountRequest(Type.ADDRESS, address, null);
+        return new AccountRequest(Type.ADDRESS, address, null, null);
     }
 
     /**
@@ -51,7 +56,17 @@ public final class AccountRequest {
      * @return the AccountRequest
      */
     public static AccountRequest from(AccountIndex index) {
-        return new AccountRequest(Type.INDEX, null, index);
+        return new AccountRequest(Type.INDEX, null, index, null);
+    }
+
+    /**
+     * Create an {@link AccountRequest} given the provided {@link CredentialRegistrationId}
+     *
+     * @param credentialRegistrationId the credential registration id.
+     * @return the AccountRequest
+     */
+    public static AccountRequest from(CredentialRegistrationId credentialRegistrationId) {
+        return new AccountRequest(Type.CREDENTIAL_REGISTRATION_ID, null, null, credentialRegistrationId);
     }
 
     @Override
@@ -61,12 +76,14 @@ public final class AccountRequest {
                 return address.encoded();
             case INDEX:
                 return index.toString();
+            case CREDENTIAL_REGISTRATION_ID:
+                return registrationId.toString();
             default:
                 throw new IllegalStateException("Invalid AccountRequest Type " + type);
         }
     }
 
     private enum Type {
-        ADDRESS, INDEX;
+        ADDRESS, INDEX, CREDENTIAL_REGISTRATION_ID;
     }
 }